Stage IV, pre-refactoring
