+++ src/main/java/com/skillbox/blogapp/model/response/TagResponseList.java
@@ -1,7 +1,7 @@
 package com.skillbox.blogapp.model.response;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.skillbox.blogapp.model.dto.TagDto;
+import com.skillbox.blogapp.service.dto.TagDto;
 import java.io.Serializable;
 import java.util.List;
 import lombok.Getter;
