+++ src/main/java/com/skillbox/blogapp/service/UserService.java
@@ -1,7 +1,7 @@
 package com.skillbox.blogapp.service;
 
-import com.skillbox.blogapp.model.dto.UserDto;
 import com.skillbox.blogapp.model.response.RegistrationResponse;
+import com.skillbox.blogapp.service.dto.UserDto;
 import java.util.List;
 import java.util.Optional;
