+++ src/main/java/com/skillbox/blogapp/service/dto/InitDto.java
@@ -1,4 +1,4 @@
-package com.skillbox.blogapp.model.dto;
+package com.skillbox.blogapp.service.dto;
 
 import lombok.Getter;
 import lombok.Setter;
