+++ src/main/java/com/skillbox/blogapp/service/dto/post/PostBriefDto.java
@@ -1,13 +1,13 @@
-package com.skillbox.blogapp.model.dto.post;
+package com.skillbox.blogapp.service.dto.post;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.skillbox.blogapp.model.dto.TagDto;
-import com.skillbox.blogapp.model.dto.UserDto;
 import com.skillbox.blogapp.model.entity.Post;
 import com.skillbox.blogapp.model.entity.PostComment;
 import com.skillbox.blogapp.model.entity.PostVote;
 import com.skillbox.blogapp.model.entity.enumeration.ModerationStatus;
+import com.skillbox.blogapp.service.dto.TagDto;
+import com.skillbox.blogapp.service.dto.UserDto;
 import java.io.Serializable;
 import java.time.Instant;
 import java.util.Set;
