+++ src/main/java/com/skillbox/blogapp/service/dto/post/PostDetailedDto.java
@@ -1,13 +1,13 @@
-package com.skillbox.blogapp.model.dto.post;
+package com.skillbox.blogapp.service.dto.post;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.skillbox.blogapp.model.dto.PostCommentDto;
-import com.skillbox.blogapp.model.dto.PostVoteDto;
-import com.skillbox.blogapp.model.dto.TagDto;
-import com.skillbox.blogapp.model.dto.UserDto;
 import com.skillbox.blogapp.model.entity.enumeration.ModerationStatus;
+import com.skillbox.blogapp.service.dto.PostCommentDto;
+import com.skillbox.blogapp.service.dto.PostVoteDto;
+import com.skillbox.blogapp.service.dto.TagDto;
+import com.skillbox.blogapp.service.dto.UserDto;
 import java.io.Serializable;
 import java.time.Instant;
 import java.util.Set;
